DRILL-6832: Removes the old "unmanaged" external sort
When the "managed" external sort was implemented a couple
of years back, we retained the original "unmanaged" version
out of an abundance of caution. The new version is now
battle tested and it is time to retire the original one.
