+++ exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/xsort/BaseWrapper.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.drill.exec.physical.impl.xsort.managed;
+package org.apache.drill.exec.physical.impl.xsort;
 
 import java.io.IOException;
 
@@ -49,5 +49,4 @@ public abstract class BaseWrapper {
             .build(logger);
     }
   }
-
 }
