+++ exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/xsort/MSortTemplate.java
@@ -17,14 +17,10 @@
  */
 package org.apache.drill.exec.physical.impl.xsort;
 
-import com.typesafe.config.ConfigException;
-import io.netty.buffer.DrillBuf;
-
 import java.util.Queue;
 
 import javax.inject.Named;
 
-import org.apache.drill.exec.ExecConstants;
 import org.apache.drill.exec.exception.SchemaChangeException;
 import org.apache.drill.exec.memory.BaseAllocator;
 import org.apache.drill.exec.memory.BufferAllocator;
@@ -32,38 +28,51 @@ import org.apache.drill.exec.ops.FragmentContext;
 import org.apache.drill.exec.record.RecordBatch;
 import org.apache.drill.exec.record.VectorContainer;
 import org.apache.drill.exec.record.selection.SelectionVector4;
-import org.apache.drill.exec.vector.ValueVector;
 import org.apache.hadoop.util.IndexedSortable;
 
 import org.apache.drill.shaded.guava.com.google.common.base.Preconditions;
 import org.apache.drill.shaded.guava.com.google.common.collect.Queues;
 
+import io.netty.buffer.DrillBuf;
+
 public abstract class MSortTemplate implements MSorter, IndexedSortable {
-//  private static final org.slf4j.Logger logger = org.slf4j.LoggerFactory.getLogger(MSortTemplate.class);
 
   private SelectionVector4 vector4;
   private SelectionVector4 aux;
+  @SuppressWarnings("unused")
+  private long compares;
+
+  /**
+   * Holds offsets into the SV4 of the start of each batch
+   * (sorted run.)
+   */
   private Queue<Integer> runStarts = Queues.newLinkedBlockingQueue();
   private FragmentContext context;
 
   /**
-   * This is only useful for debugging and/or unit testing. Controls the maximum size of batches exposed to downstream
+   * Controls the maximum size of batches exposed to downstream
    */
   private int desiredRecordBatchCount;
 
   @Override
-  public void setup(final FragmentContext context, final BufferAllocator allocator, final SelectionVector4 vector4, final VectorContainer hyperBatch) throws SchemaChangeException{
+  public void setup(FragmentContext context, BufferAllocator allocator, SelectionVector4 vector4,
+                    VectorContainer hyperBatch, int outputBatchSize, int desiredBatchSize) throws SchemaChangeException{
     // we pass in the local hyperBatch since that is where we'll be reading data.
     Preconditions.checkNotNull(vector4);
     this.vector4 = vector4.createNewWrapperCurrent();
     this.context = context;
     vector4.clear();
     doSetup(context, hyperBatch, null);
+
+    // Populate the queue with the offset in the SV4 of each
+    // batch. Note that this is expensive as it requires a scan
+    // of all items to be sorted: potentially millions.
+
     runStarts.add(0);
     int batch = 0;
-    final int totalCount = this.vector4.getTotalCount();
+    int totalCount = this.vector4.getTotalCount();
     for (int i = 0; i < totalCount; i++) {
-      final int newBatch = this.vector4.get(i) >>> 16;
+      int newBatch = this.vector4.get(i) >>> 16;
       if (newBatch == batch) {
         continue;
       } else if (newBatch == batch + 1) {
@@ -73,31 +82,34 @@ public abstract class MSortTemplate implements MSorter, IndexedSortable {
         throw new UnsupportedOperationException(String.format("Missing batch. batch: %d newBatch: %d", batch, newBatch));
       }
     }
-    final DrillBuf drillBuf = allocator.buffer(4 * totalCount);
 
-    try {
-      desiredRecordBatchCount = context.getConfig().getInt(ExecConstants.EXTERNAL_SORT_MSORT_MAX_BATCHSIZE);
-    } catch(ConfigException.Missing e) {
-      // value not found, use default value instead
-      desiredRecordBatchCount = ValueVector.MAX_ROW_COUNT;
-    }
+    // Create a temporary SV4 to hold the merged results.
+
+    DrillBuf drillBuf = allocator.buffer(4 * totalCount);
+    desiredRecordBatchCount = Math.min(outputBatchSize, desiredBatchSize);
+    desiredRecordBatchCount = Math.min(desiredRecordBatchCount, totalCount);
     aux = new SelectionVector4(drillBuf, totalCount, desiredRecordBatchCount);
   }
 
   /**
-   * For given recordCount how much memory does MSorter needs for its own purpose. This is used in
-   * ExternalSortBatch to make decisions about whether to spill or not.
-   *
-   * @param recordCount
-   * @return The amount of memory MSorter needs for a given record count.
+   * For given recordCount how much memory does MSorter needs for its own
+   * purpose. This is used in ExternalSortBatch to make decisions about whether
+   * to spill or not.
    */
-  public static long memoryNeeded(final int recordCount) {
-    // We need 4 bytes (SV4) for each record, power of 2 rounded.
+  public static long memoryNeeded(int recordCount) {
+    // We need 4 bytes (SV4) for each record.
+    // The memory allocator will round this to the next
+    // power of 2.
 
     return BaseAllocator.nextPowerOfTwo(recordCount * 4);
   }
 
-  private int merge(final int leftStart, final int rightStart, final int rightEnd, final int outStart) {
+  /**
+   * Given two regions within the selection vector 4 (a left and a right), merge
+   * the two regions to produce a combined output region in the auxiliary
+   * selection vector.
+   */
+  protected int merge(int leftStart, int rightStart, int rightEnd, int outStart) {
     int l = leftStart;
     int r = rightStart;
     int o = outStart;
@@ -123,35 +135,45 @@ public abstract class MSortTemplate implements MSorter, IndexedSortable {
     return vector4;
   }
 
+  /**
+   * Merge a set of pre-sorted runs to produce a combined
+   * result set. Merging is done in the selection vector, record data does
+   * not move.
+   * <p>
+   * Runs are merge pairwise in multiple passes, providing performance
+   * of O(n * m * log(n)), where n = number of runs, m = number of records
+   * per run.
+   */
   @Override
-  public void sort(final VectorContainer container) {
+  public void sort() {
     while (runStarts.size() > 1) {
+      int totalCount = vector4.getTotalCount();
 
-      // check if we're cancelled/failed frequently
+      // check if we're cancelled/failed recently
       if (!context.getExecutorState().shouldContinue()) {
         return;
       }
 
       int outIndex = 0;
-      final Queue<Integer> newRunStarts = Queues.newLinkedBlockingQueue();
+      Queue<Integer> newRunStarts = Queues.newLinkedBlockingQueue();
       newRunStarts.add(outIndex);
-      final int size = runStarts.size();
+      int size = runStarts.size();
       for (int i = 0; i < size / 2; i++) {
-        final int left = runStarts.poll();
-        final int right = runStarts.poll();
+        int left = runStarts.poll();
+        int right = runStarts.poll();
         Integer end = runStarts.peek();
         if (end == null) {
-          end = vector4.getTotalCount();
+          end = totalCount;
         }
         outIndex = merge(left, right, end, outIndex);
         if (outIndex < vector4.getTotalCount()) {
           newRunStarts.add(outIndex);
         }
       }
-      if (outIndex < vector4.getTotalCount()) {
-        copyRun(outIndex, vector4.getTotalCount());
+      if (outIndex < totalCount) {
+        copyRun(outIndex, totalCount);
       }
-      final SelectionVector4 tmp = aux.createNewWrapperCurrent(desiredRecordBatchCount);
+      SelectionVector4 tmp = aux.createNewWrapperCurrent(desiredRecordBatchCount);
       aux.clear();
       aux = vector4.createNewWrapperCurrent(desiredRecordBatchCount);
       vector4.clear();
@@ -162,38 +184,40 @@ public abstract class MSortTemplate implements MSorter, IndexedSortable {
     aux.clear();
   }
 
-  private void copyRun(final int start, final int end) {
+  private void copyRun(int start, int end) {
     for (int i = start; i < end; i++) {
       aux.set(i, vector4.get(i));
     }
   }
 
   @Override
-  public void swap(final int sv0, final int sv1) {
-    final int tmp = vector4.get(sv0);
+  public void swap(int sv0, int sv1) {
+    int tmp = vector4.get(sv0);
     vector4.set(sv0, vector4.get(sv1));
     vector4.set(sv1, tmp);
   }
 
   @Override
-  public int compare(final int leftIndex, final int rightIndex) {
-    final int sv1 = vector4.get(leftIndex);
-    final int sv2 = vector4.get(rightIndex);
+  public int compare(int leftIndex, int rightIndex) {
+    int sv1 = vector4.get(leftIndex);
+    int sv2 = vector4.get(rightIndex);
+    compares++;
     try {
       return doEval(sv1, sv2);
     } catch (SchemaChangeException e) {
-      throw new UnsupportedOperationException(e);
+      throw new IllegalStateException(e);
     }
   }
 
   @Override
   public void clear() {
-    if(vector4 != null) {
+    if (vector4 != null) {
       vector4.clear();
+      vector4 = null;
     }
-
-    if(aux != null) {
+    if (aux != null) {
       aux.clear();
+      aux = null;
     }
   }
 
@@ -204,13 +228,4 @@ public abstract class MSortTemplate implements MSorter, IndexedSortable {
   public abstract int doEval(@Named("leftIndex") int leftIndex,
                              @Named("rightIndex") int rightIndex)
                       throws SchemaChangeException;
-
-  @Override
-  public String toString() {
-    return "MSortTemplate[vector4=" + vector4
-        + ", aux=" + aux
-        + ", runStarts=" + runStarts
-        + ", desiredRecordBatchCount=" + desiredRecordBatchCount
-        + "]";
-  }
 }
