+++ exec/java-exec/src/main/java/org/apache/drill/exec/physical/impl/xsort/PriorityQueueCopierTemplate.java
@@ -17,8 +17,6 @@
  */
 package org.apache.drill.exec.physical.impl.xsort;
 
-import io.netty.buffer.DrillBuf;
-
 import java.io.IOException;
 import java.util.List;
 
@@ -26,14 +24,13 @@ import javax.inject.Named;
 
 import org.apache.drill.exec.exception.SchemaChangeException;
 import org.apache.drill.exec.memory.BufferAllocator;
-import org.apache.drill.exec.ops.FragmentContext;
 import org.apache.drill.exec.record.VectorAccessible;
 import org.apache.drill.exec.record.VectorAccessibleUtilities;
 import org.apache.drill.exec.record.selection.SelectionVector4;
 
+import io.netty.buffer.DrillBuf;
 
 public abstract class PriorityQueueCopierTemplate implements PriorityQueueCopier {
-
   private SelectionVector4 vector4;
   private List<BatchGroup> batchGroups;
   private VectorAccessible hyperBatch;
@@ -42,9 +39,8 @@ public abstract class PriorityQueueCopierTemplate implements PriorityQueueCopier
   private int queueSize;
 
   @Override
-  public void setup(FragmentContext context, BufferAllocator allocator,
-      VectorAccessible hyperBatch, List<BatchGroup> batchGroups,
-      VectorAccessible outgoing) throws SchemaChangeException {
+  public void setup(BufferAllocator allocator, VectorAccessible hyperBatch, List<BatchGroup> batchGroups,
+                    VectorAccessible outgoing) throws SchemaChangeException {
     this.hyperBatch = hyperBatch;
     this.batchGroups = batchGroups;
     this.outgoing = outgoing;
@@ -52,45 +48,47 @@ public abstract class PriorityQueueCopierTemplate implements PriorityQueueCopier
 
     final DrillBuf drillBuf = allocator.buffer(4 * size);
     vector4 = new SelectionVector4(drillBuf, size, Character.MAX_VALUE);
-    doSetup(context, hyperBatch, outgoing);
+    doSetup(hyperBatch, outgoing);
 
     queueSize = 0;
     for (int i = 0; i < size; i++) {
-      vector4.set(i, i, batchGroups.get(i).getNextIndex());
-      siftUp();
-      queueSize++;
+      int index = batchGroups.get(i).getNextIndex();
+      vector4.set(i, i, index);
+      if (index > -1) {
+        siftUp();
+        queueSize++;
+      }
     }
   }
 
   @Override
   public int next(int targetRecordCount) {
-    allocateVectors(targetRecordCount);
-    for (int outgoingIndex = 0; outgoingIndex < targetRecordCount; outgoingIndex++) {
-      if (queueSize == 0) {
-        return 0;
+    try {
+     for (int outgoingIndex = 0; outgoingIndex < targetRecordCount; outgoingIndex++) {
+        if (queueSize == 0) {
+          return 0;
+        }
+        int compoundIndex = vector4.get(0);
+        int batch = compoundIndex >>> 16;
+        assert batch < batchGroups.size() : String.format("batch: %d batchGroups: %d", batch, batchGroups.size());
+        doCopy(compoundIndex, outgoingIndex);
+        int nextIndex = batchGroups.get(batch).getNextIndex();
+        if (nextIndex < 0) {
+          vector4.set(0, vector4.get(--queueSize));
+        } else {
+          vector4.set(0, batch, nextIndex);
+        }
+        if (queueSize == 0) {
+          VectorAccessibleUtilities.setValueCount(outgoing, ++outgoingIndex);
+          return outgoingIndex;
+        }
+        siftDown();
       }
-      int compoundIndex = vector4.get(0);
-      int batch = compoundIndex >>> 16;
-      assert batch < batchGroups.size() : String.format("batch: %d batchGroups: %d", batch, batchGroups.size());
-      doCopy(compoundIndex, outgoingIndex);
-      int nextIndex = batchGroups.get(batch).getNextIndex();
-      if (nextIndex < 0) {
-        vector4.set(0, vector4.get(--queueSize));
-      } else {
-        vector4.set(0, batch, nextIndex);
-      }
-      if (queueSize == 0) {
-        setValueCount(++outgoingIndex);
-        return outgoingIndex;
-      }
-      siftDown();
+      VectorAccessibleUtilities.setValueCount(outgoing, targetRecordCount);
+      return targetRecordCount;
+    } catch (SchemaChangeException e) {
+      throw new IllegalStateException(e);
     }
-    setValueCount(targetRecordCount);
-    return targetRecordCount;
-  }
-
-  private void setValueCount(int count) {
-    VectorAccessibleUtilities.setValueCount(outgoing, count);
   }
 
   @Override
@@ -98,13 +96,10 @@ public abstract class PriorityQueueCopierTemplate implements PriorityQueueCopier
     vector4.clear();
     VectorAccessibleUtilities.clear(outgoing);
     VectorAccessibleUtilities.clear(hyperBatch);
-
-    for (BatchGroup batchGroup : batchGroups) {
-      batchGroup.close();
-    }
+    BatchGroup.closeAll(batchGroups);
   }
 
-  private void siftUp() {
+  private void siftUp() throws SchemaChangeException {
     int p = queueSize;
     while (p > 0) {
       if (compare(p, (p - 1) / 2) < 0) {
@@ -116,11 +111,7 @@ public abstract class PriorityQueueCopierTemplate implements PriorityQueueCopier
     }
   }
 
-  private void allocateVectors(int targetRecordCount) {
-    VectorAccessibleUtilities.allocateVectors(outgoing, targetRecordCount);
-  }
-
-  private void siftDown() {
+  private void siftDown() throws SchemaChangeException {
     int p = 0;
     int next;
     while (p * 2 + 1 < queueSize) { // While the current node has at least one child
@@ -148,14 +139,19 @@ public abstract class PriorityQueueCopierTemplate implements PriorityQueueCopier
     vector4.set(sv1, tmp);
   }
 
-  public int compare(int leftIndex, int rightIndex) {
+  public int compare(int leftIndex, int rightIndex) throws SchemaChangeException {
     int sv1 = vector4.get(leftIndex);
     int sv2 = vector4.get(rightIndex);
     return doEval(sv1, sv2);
   }
 
-  public abstract void doSetup(@Named("context") FragmentContext context,
-      @Named("incoming") VectorAccessible incoming, @Named("outgoing") VectorAccessible outgoing);
-  public abstract int doEval(@Named("leftIndex") int leftIndex, @Named("rightIndex") int rightIndex);
-  public abstract void doCopy(@Named("inIndex") int inIndex, @Named("outIndex") int outIndex);
+  public abstract void doSetup(@Named("incoming") VectorAccessible incoming,
+                               @Named("outgoing") VectorAccessible outgoing)
+                       throws SchemaChangeException;
+  public abstract int doEval(@Named("leftIndex") int leftIndex,
+                             @Named("rightIndex") int rightIndex)
+                      throws SchemaChangeException;
+  public abstract void doCopy(@Named("inIndex") int inIndex,
+                              @Named("outIndex") int outIndex)
+                       throws SchemaChangeException;
 }
