+++ exec/java-exec/src/main/resources/drill-module.conf
@@ -531,8 +531,11 @@ drill.exec.options: {
     exec.queue.timeout_millis: 300000,
     exec.queue.memory_ratio: 10.0,
     exec.queue.memory_reserve_ratio: 0.2,
+    # Deprecated
     exec.sort.disable_managed : false,
+    # Deprecated
     exec.storage.enable_new_text_reader: true,
+    # Deprecated
     exec.storage.enable_v3_text_reader: false,
     exec.storage.min_width: 1,
     exec.storage.skip_runtime_rowgroup_pruning: false,
